Add Activities events controller. update event show page to correctly show activity information.
