+++ app/models/activities_event.rb
@@ -3,7 +3,7 @@ class ActivitiesEvent < ApplicationRecord
   belongs_to :event
 
   def display_title
-    custom_title.presence || activity.title
+    custom_title.presence || activity.title.capitalize
   end
 
   def display_description
