+++ app/views/activities_events/_activity_form.html.erb
@@ -7,7 +7,7 @@
     <%= f.input :custom_description, label: "Activity Description", input_html: { class: "form-control" } %>
   </div>
 
-  <div class="d-grid">
+  <div class="">
     <%= f.button :submit, @activity_event.new_record? ? "Add Activity" : "Update Activity", class: "btn btn-primary btn-lg" %>
   </div>
 <% end %>
