+++ app/views/activities_events/new.html.erb
@@ -1,0 +1,3 @@
+<div class="container my-4">
+  <%= render "activities_events/activity_form", event: @event, activity_event: @activity_event %>
+</div>
