--- app/views/events/new_activity.html.erb
@@ -1,1 +1,0 @@
-<%= render "events/activity_form", event: @event, activity_event: @activity_event %>
