+++ config/routes.rb
@@ -17,8 +17,8 @@ Rails.application.routes.draw do
   get "invite_link", to: "invites#invite_link"
   get "organizations/join", to: "invites#join"
   post "organizations/invite_result", to: "invites#invite_result"
-  get "events/:id/new_activity", to: "events#new_activity", as: :new_activity
-  post "events/:id/new_activity", to: "events#create_activity", as: :create_activity
+  get "events/:event_id/activities/new", to: "activities_events#new", as: :new_activity
+  post "events/:event_id/activities", to: "activities_events#create", as: :create_activity
 
   resources :organizations, only: [:show, :index, :new, :create] do
     get :invite, on: :member
